requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy
pandas
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
six==1.15.0
twilio==6.52.0
urllib3==1.26.3
Werkzeug==1.0.1
gunicorn
